+++ ciroh_plugins/drought/sourceUrls.py
@@ -81,7 +81,7 @@ ImageTile_urls = {
     "vegdri_url_esri": "https://tiles.arcgis.com/tiles/0OTVzJS4K09zlixn/arcgis/rest/services/VegDRI_Example/MapServer",
     "land_cover_url_esri": "https://tiles.arcgis.com/tiles/0OTVzJS4K09zlixn/arcgis/rest/services/Land_Cover_2020/MapServer/",
     "land_use_url_esri": "https://tiles.arcgis.com/tiles/0OTVzJS4K09zlixn/arcgis/rest/services/Land_Use_2020/MapServer/",
-    "biomass_example_url_esri": "https://tiles.arcgis.com/tiles/0OTVzJS4K09zlixn/arcgis/rest/services/Biomass_2018_masked_tif/MapServer/",
+    "biomass_example_url_esri": "https://tiles.arcgis.com/tiles/0OTVzJS4K09zlixn/arcgis/rest/services/Biomass_2018_masked_tif/MapServer/",  # not show up
     "huc4_tile_url_esri": "https://tiles.arcgis.com/tiles/0OTVzJS4K09zlixn/arcgis/rest/services/HUC4_Simplified/MapServer/",
     "county_tile_url_esri": "https://tiles.arcgis.com/tiles/0OTVzJS4K09zlixn/arcgis/rest/services/Counties_Simplified/MapServer/",
     "nws_radar_tile_url_esri": "https://tiles.arcgis.com/tiles/0OTVzJS4K09zlixn/arcgis/rest/services/NWS_Radar_10k/MapServer",
