+++ pyproject.toml
@@ -41,8 +41,8 @@ nwmp_gauges_series = "ciroh_plugins.nwmps.gauges:NWMPSGaugesSeries"
 drought_drought_monitoring_table = "ciroh_plugins.drought.drought_table:DroughtDataTable"
 drought_drought_monitoring_series = "ciroh_plugins.drought.drought_plot:DroughtDataTimeSeries"
 drought_drought_monitoring_pie_chart = "ciroh_plugins.drought.drought_pie:DroughtDataGraph"
-drought_drought_map = "ciroh_plugins.drought.map:Map"
-drought_map_layer = "ciroh_plugins.drought.layer_finder:LayerFinder"
+drought_map = "ciroh_plugins.drought.drought_map:DroughtMap"
+drought_map_layer = "ciroh_plugins.drought.drought_map_layer_finder:DroughtMapLayerFinder"
 
 [tool.setuptools]
 include-package-data = true
